+++ devolo_home_control_api/devices/zwave.py
@@ -8,7 +8,7 @@ class Zwave:
     :param device_uid: Device UID, something like hdm:ZWave:CBC56091/24
     """
 
-    def __init__(self, name, device_uid, zone, battery_level, icon):
+    def __init__(self, name: str, device_uid: str, zone: str, battery_level: int, icon: str, online_state: int):
         self._logger = logging.getLogger(self.__class__.__name__)
         self.name = name
         self.zone = zone
@@ -17,3 +17,22 @@ class Zwave:
         self.icon = icon
         self.device_uid = device_uid
         self.subscriber = None
+
+        # Online state is returned as numbers. 1 --> Offline, 2 --> Online, 7 (?) --> Not initialized
+        if online_state == 1:
+            self.online = "offline"
+        elif online_state == 2:
+            self.online = "online"
+        else:
+            self.online = "unknown state"
+            self._logger.warning(f"Unknown state {online_state} for device {self.name}")
+
+    def get_property(self, name: str) -> list:
+        """
+        Get element UIDs to a specified property.
+
+        :param name: Name of the property we want to access
+        :return: List of UIDs in this property
+        :raises: AttributeError: The property does not exist in this device type
+        """
+        return [*getattr(self, f"{name}_property").keys()]