+++ tests/conftest.py
@@ -1,15 +1,21 @@
 import json
+import sys
 
 import pytest
 
 from devolo_home_control_api.mprm_rest import MprmRest
 from devolo_home_control_api.mprm_websocket import MprmWebsocket
 from devolo_home_control_api.mydevolo import Mydevolo, WrongUrlError
+from tests.mock_dummy_device import dummy_device
 from tests.mock_gateway import Gateway
 from tests.mock_metering_plug import metering_plug
 
-with open('test_data.json') as file:
-    test_data = json.load(file)
+try:
+    with open('test_data.json') as file:
+        test_data = json.load(file)
+except FileNotFoundError:
+    print("Please run tests from within the tests directory.")
+    sys.exit(127)
 
 
 @pytest.fixture()
@@ -20,8 +26,12 @@ def mock_gateway(mocker):
 @pytest.fixture()
 def mock_inspect_devices_metering_plug(mocker):
     def mock__inspect_devices(self):
-        device_uid = test_data.get("device").get("mains").get("uid")
-        self.devices[device_uid] = metering_plug(device_uid=device_uid)
+        for device_type, device in test_data.get("devices").items():
+            device_uid = device.get("uid")
+            if device_type == "mains":
+                self.devices[device_uid] = metering_plug(device_uid=device_uid)
+            else:
+                self.devices[device_uid] = dummy_device(key=device_type)
 
     mocker.patch("devolo_home_control_api.mprm_rest.MprmRest._inspect_devices", mock__inspect_devices)
 
@@ -51,7 +61,14 @@ def mock_mydevolo__call(mocker, request):
             else:
                 return {"items": [{"gatewayId": gateway_id}]}
         elif url == f"https://www.mydevolo.com/v1/users/{uuid}/hc/gateways/{gateway_id}":
-            return {"gatewayId": gateway_id}
+            return {"gatewayId": gateway_id,
+                    "status": "devolo.hc_gateway.status.online",
+                    "state": "devolo.hc_gateway.state.idle"}
+        elif url == "https://www.mydevolo.com/v1/hc/maintenance":
+            if request.node.name == "test_maintenance_off":
+                return {"state": "off"}
+            else:
+                return {"state": "on"}
         else:
             raise WrongUrlError(f"This URL was not mocked: {url}")
 
@@ -61,30 +78,32 @@ def mock_mydevolo__call(mocker, request):
 @pytest.fixture()
 def mock_mprmrest__extract_data_from_element_uid(mocker, request):
     def _extract_data_from_element_uid(element_uid):
-        if request.node.name == "test_get_binary_switch_state_valid_on":
-            return {"properties": {"state": 1}}
-        elif request.node.name == "test_get_binary_switch_state_valid_off":
-            return {"properties": {"state": 0}}
-        elif request.node.name == "test_get_consumption_valid":
-            return {"properties": {"currentValue": test_data.get("device").get("mains").get("current_consumption"),
-                                   "totalValue": test_data.get("device").get("mains").get("total_consumption")}}
-        elif request.node.name == "test_get_led_setting_valid":
-            return {"properties": {"led": test_data.get("device").get("mains").get("led_setting")}}
-        elif request.node.name == "test_get_param_changed_valid":
-            return {"properties": {"paramChanged": test_data.get("device").get("mains").get("param_changed")}}
-        elif request.node.name == "test_get_general_device_settings_valid":
-            return {"properties": {"eventsEnabled": test_data.get("device").get("mains").get("events_enabled"),
-                                   "name": test_data.get("device").get("mains").get("name"),
-                                   "icon": test_data.get("device").get("mains").get("icon"),
-                                   "zoneID": test_data.get("device").get("mains").get("zone_id")}}
-        elif request.node.name == "test_get_protection_setting_valid":
-            return {"properties": {"localSwitch": test_data.get("device").get("mains").get("local_switch"),
-                                   "remoteSwitch": test_data.get("device").get("mains").get("remote_switch")}}
-        elif request.node.name == "test_get_voltage_valid":
-            return {"properties": {"value": test_data.get("device").get("mains").get("voltage")}}
-        elif request.node.name == "test_update_consumption_valid":
-            return {"properties": {"currentValue": test_data.get("device").get("mains").get("current_consumption"),
-                                   "totalValue": test_data.get("device").get("mains").get("total_consumption")}}
+        properties = {}
+        properties['test_get_binary_switch_state_valid_on'] = {
+            "properties": {"state": 1}}
+        properties['test_get_binary_switch_state_valid_off'] = {
+            "properties": {"state": 0}}
+        properties['test_get_consumption_valid'] = {
+            "properties": {"currentValue": test_data.get("devices").get("mains").get("current_consumption"),
+                           "totalValue": test_data.get("devices").get("mains").get("total_consumption")}}
+        properties['test_get_led_setting_valid'] = {
+            "properties": {"led": test_data.get("devices").get("mains").get("led_setting")}}
+        properties['test_get_param_changed_valid'] = {
+            "properties": {"paramChanged": test_data.get("devices").get("mains").get("param_changed")}}
+        properties['test_get_general_device_settings_valid'] = {
+            "properties": {"eventsEnabled": test_data.get("devices").get("mains").get("events_enabled"),
+                           "name": test_data.get("devices").get("mains").get("name"),
+                           "icon": test_data.get("devices").get("mains").get("icon"),
+                           "zoneID": test_data.get("devices").get("mains").get("zone_id")}}
+        properties['test_get_protection_setting_valid'] = {
+            "properties": {"localSwitch": test_data.get("devices").get("mains").get("local_switch"),
+                           "remoteSwitch": test_data.get("devices").get("mains").get("remote_switch")}}
+        properties['test_get_voltage_valid'] = {
+            "properties": {"value": test_data.get("devices").get("mains").get("voltage")}}
+        properties['test_update_consumption_valid'] = {
+            "properties": {"currentValue": test_data.get("devices").get("mains").get("current_consumption"),
+                           "totalValue": test_data.get("devices").get("mains").get("total_consumption")}}
+        return properties.get(request.node.name)
 
     mocker.patch("devolo_home_control_api.mprm_rest.MprmRest._extract_data_from_element_uid",
                  side_effect=_extract_data_from_element_uid)
@@ -96,7 +115,7 @@ def mock_mprmrest__post_set(mocker, request):
         if request.node.name == "test_set_binary_switch_valid":
             return {"result": {"status": 1}}
         elif request.node.name == "test_set_binary_switch_error":
-            return {"result": {"status": 3}}
+            return {"result": {"status": 2}}
 
     mocker.patch("devolo_home_control_api.mprm_rest.MprmRest._post", side_effect=_post_mock)
 
@@ -105,7 +124,7 @@ def mock_mprmrest__post_set(mocker, request):
 def test_data_fixture(request):
     request.cls.user = test_data.get("user")
     request.cls.gateway = test_data.get("gateway")
-    request.cls.device = test_data.get("device")
+    request.cls.devices = test_data.get("devices")
 
 
 @pytest.fixture()
@@ -115,6 +134,7 @@ def mprm_instance(request, mocker, mock_gateway, mock_inspect_devices_metering_p
     else:
         def _websocket_connection_mock():
             pass
+
         mocker.patch("devolo_home_control_api.mprm_websocket.MprmWebsocket._websocket_connection",
                      side_effect=_websocket_connection_mock)
         request.cls.mprm = MprmWebsocket(test_data.get("gateway").get("id"))
@@ -122,15 +142,17 @@ def mprm_instance(request, mocker, mock_gateway, mock_inspect_devices_metering_p
 
 @pytest.fixture()
 def fill_device_data(request):
-    consumption_property = request.cls.mprm.devices.get(test_data.get('device').get("mains").get("uid")).consumption_property
-    consumption_property.get(f"devolo.Meter:{request.cls.device.get('mains').get('uid')}").current = 0.58
-    consumption_property.get(f"devolo.Meter:{request.cls.device.get('mains').get('uid')}").total = 125.68
+    consumption_property = request.cls.mprm.devices.get(test_data.get('devices').get("mains").get("uid")).consumption_property
+    consumption_property.get(f"devolo.Meter:{test_data.get('devices').get('mains').get('uid')}").current = 0.58
+    consumption_property.get(f"devolo.Meter:{test_data.get('devices').get('mains').get('uid')}").total = 125.68
 
-    binary_switch_property = request.cls.mprm.devices.get(test_data.get('device').get("mains").get("uid")).binary_switch_property
-    binary_switch_property.get(f"devolo.BinarySwitch:{request.cls.device.get('mains').get('uid')}").state = False
+    binary_switch_property = \
+        request.cls.mprm.devices.get(test_data.get('devices').get("mains").get("uid")).binary_switch_property
+    binary_switch_property.get(f"devolo.BinarySwitch:{test_data.get('devices').get('mains').get('uid')}").state = False
 
-    voltage_property = request.cls.mprm.devices.get(test_data.get('device').get("mains").get("uid")).voltage_property
-    voltage_property.get(f"devolo.VoltageMultiLevelSensor:{request.cls.device.get('mains').get('uid')}").current = 236
+    voltage_property = request.cls.mprm.devices.get(test_data.get('devices').get("mains").get("uid")).voltage_property
+    voltage_property.get(f"devolo.VoltageMultiLevelSensor:{test_data.get('devices').get('mains').get('uid')}").current = 236
+
 
 @pytest.fixture(autouse=True)
 def clear_mydevolo():