+++ docs/CHANGELOG.md
@@ -1,10 +1,27 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v0.1.0]
+## [v0.2.0] - 05.02.2020
+
 ### Added
+
+- Lookup methods, if UIDs are unknown
+- React on maintenance mode of mydevolo
+- Handle offline gateways
+- Handle offline devices
+
+### Changed
+
+- Publisher now returns a list of element_uid and value
+- Rename MprmDeviceError to MprmDeviceCommunicationError
+
+## [v0.1.0] - 31.01.2020
+
+### Added
+
 - Support for devolo Metering Plug v1
 - Support for devolo Metering Plug v2
 - Support for devolo Switch FM